requirements.txt
flask
flask-cors
python-dotenv
requests
urllib3